[struct_pack] refactor code
