+++ include/ylt/struct_pack/util.h
@@ -18,6 +18,10 @@
 
 #include "marco.h"
 namespace struct_pack::detail {
+
+template <typename T>
+using remove_cvref_t = std::remove_cv_t<std::remove_reference_t<T>>;
+
 [[noreturn]] inline void unreachable() {
   // Uses compiler specific extensions if possible.
   // Even if no extension is used, undefined behavior is still raised by
