+++ include/ylt/struct_pack/varint.hpp
@@ -1,3 +1,18 @@
+/*
+ * Copyright (c) 2023, Alibaba Group Holding Limited;
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 #pragma once
 #include <cstdint>
 #include <ostream>
@@ -256,5 +271,8 @@ template <bool NotSkip = true,
   // return decode_varint_v1(f);
 }
 }  // namespace detail
-
+using var_int32_t = detail::sint<int32_t>;
+using var_int64_t = detail::sint<int64_t>;
+using var_uint32_t = detail::varint<uint32_t>;
+using var_uint64_t = detail::varint<uint64_t>;
 }  // namespace struct_pack
